Author: Philip Craig Date: 2016-08-04 16:50:57 Revision: d4ab3c25edc32c50e06e98531aa50c787fde7645 Comment: netifd: fix seg fault for dhcp on bridge
device_init_settings() must always be called using device_attr_list.
Prior to this change, we were using the attributes specific to the
type of device, which is wrong, and wasn't apparent until we used
a bridge as the device.

dhcp is the only protocol that uses this code path (it uses it to set
the mtu).