Author: Philip Craig Date: 2015-05-04 15:13:09 Revision: 58c72b3828a1ea13909f3cd7cc5ceb8a0db6dfbf Comment: netifd: import new version
* user/netifd/patches/netifd-interface-cleanup-l3-dev.patch
Fixed in mainline commit 77206574.

* user/netifd/patches/netifd-link-status.patch
A different version of this patch is now in mainline.

* user/netifd/patches/netifd-lose-def-route-with-metric.patch
Fixed in mainline commit ceeed831.

* user/netifd/patches/netifd-proto-route-mtu.patch
No longer applies cleanly, and we don't need it.

* user/netifd/patches/netifd-set-subnet-route-src-when-using-metric.patch
Simplified due to mainline commit 25289c94.

* user/netifd/patches/netifd-xxx-route-preferred-source.patch
Replaced by mainline commit 25289c94. This requires us to use the
"source" keyword in the uci config, instead of "prefsrc".