Author: Philip Craig Date: 2015-06-26 16:43:44 Revision: 0a398e1370fd88f1895e2dcdb9f5f297e9156ac5 Comment: netifd: fix NULL device on network restart
This is a workaround for a netifd bug. It occurs if there are two
or more static addresses on the same network device, and you run
'ubus call network restart'.

The problem is that the second interface is brought up as a result of
the first interface coming up, but the code path that does this doesn't
claim the device. So when the interfaces are taken down, the device is
released before the second interface is down, and then when the
second interface is taken down, system_del_address() is called with
a NULL dev.