Author: Samuel Date: 2017-11-07 11:16:53 Revision: 0a71bb302c38f4110b605243127220316cd79013 Comment: user/netifd: updated netifd-link-loss-dynamic.patch to only apply to dynamic interfaces
The patch is only required to ensure child interfaces of the modem (i.e.
ipv4_interface_modem and ipv6_interface_modem) are cleaned up after the
modem device disappears. So only apply the "fix" to dynamic interfaces
(which the child interfaces are).

This change is required because applying it to static interfaces causes issues
when force_link is disabled.