Author: Brendan Chen Date: 2017-12-07 17:20:19 Revision: 341612c68f6ff803e54386bd7c877531f194db39 Comment: netifd/patches: Fixed failover issue when link disappears quickly
Failover did not work when the primary link breaks after a reconnecting
for less than 5 seconds. Firewall action did not kick in when
disconnection happened under the above conditions and that was due to
the action event not kicking in.

This is now fixed by adding the action event trigger when IFPEV and IFS
are both in the _DOWN state.