Author: Samuel Brian Date: 2020-02-21 18:23:42 Revision: 02597cc9ef6fab80ab39fc3c8381a55850cc80b5 Comment: user/netifd: set config reload flag in proto setup
Set an environment variable when the proto shell script is being run
because of a "reload" event.
This can be observed in the proto "setup" after a configuration change
is made to the interface.
When the variable is not present, one can assume that an down/up event
is responsible for the setup.

Determining reload vs down/up is required for the passthrough.sh proto
to decide whether to toggle link on the underlying network device.

DAL-2642