Author: jps Date: 2020-11-06 19:17:58 Revision: 6df390acdcac919fce063e6c186007304ac7cb65 Comment: netifd: in rare cases, the bridge asserts if the wlan macaddr changes
If the wlan macaddr changes (eg. from the default to the configured
value) whilst being added to the bridge, it would assert. This was the
only assert left in the bridge code from the original 2011 creation. All
other routines simply checked and returned. We now do this.

Tested on an EX15W and IX20W

DAL-3949